shared state restored
